+++ src/layout/sidebar.component.ts
@@ -5,6 +5,7 @@ import {
     EventEmitter,
     Host,
     HostBinding,
+    HostListener,
     Input,
     OnDestroy,
     OnInit,
@@ -46,14 +47,14 @@ export type ThySidebarTheme = 'white' | 'light' | 'dark';
             (thyResize)="resizeHandler($event)"
             (thyResizeStart)="resizeStart()"
             (thyResizeEnd)="resizeEnd()"
-            [style.display]="!collapseVisible ? 'contents' : null">
+            [style.display]="!isResizable ? 'contents' : null">
             <thy-resize-handle
                 *ngIf="!thyCollapsed"
                 thyDirection="right"
                 class="sidebar-resize-handle"
                 thyLine="true"
-                (mouseenter)="resizeHandleHover($event, 'enter')"
-                (mouseleave)="resizeHandleHover($event, 'leave')"
+                (mouseenter)="toggleResizable($event, 'enter')"
+                (mouseleave)="toggleResizable($event, 'leave')"
                 (dblclick)="restoreToDefaultWidth()">
             </thy-resize-handle>
         </div>
@@ -102,6 +103,16 @@ export class ThySidebarComponent implements OnInit, OnDestroy {
     }
 
     @HostBinding('class.thy-layout-sidebar-isolated') sidebarIsolated = false;
+
+    @HostListener('mouseenter', ['$event'])
+    mouseenter($event: MouseEvent) {
+        this.resizeHandleHover($event, 'enter');
+    }
+
+    @HostListener('mouseleave', ['$event'])
+    mouseleave($event: MouseEvent) {
+        this.resizeHandleHover($event, 'leave');
+    }
 
     /**
      * 宽度, 默认是 240px, 传入 lg 大小时宽度是300px
@@ -238,6 +249,8 @@ export class ThySidebarComponent implements OnInit, OnDestroy {
 
     isRemoveTransition: boolean;
 
+    isResizable: boolean;
+
     private hotkeySubscription: Subscription;
 
     constructor(
@@ -295,13 +308,17 @@ export class ThySidebarComponent implements OnInit, OnDestroy {
     }
 
     resizeHandleHover(event: MouseEvent, type: 'enter' | 'leave') {
-        this.collapseVisible = type === 'enter' ? true : false;
+        this.collapseVisible = type === 'enter' && !this.thyCollapsed ? true : false;
     }
 
     toggleCollapse(event?: MouseEvent) {
         this.thyCollapsed = !this.thyCollapsed;
         setTimeout(() => this.updateCollapseTip(), 200);
         this.thyCollapsedChange.emit(this.isCollapsed);
+    }
+
+    public toggleResizable(event: MouseEvent, type: 'enter' | 'leave') {
+        this.isResizable = type === 'enter' ? true : false;
     }
 
     restoreToDefaultWidth() {