+++ src/tree/examples/mocks.ts
@@ -120,7 +120,24 @@ export const treeNodes: any[] = [
                         parent_id: '5d4e5b365fadf30311c3d889',
                         position: 1114112,
                         member_count: 0,
-                        children: []
+                        disabled: true,
+                        children: [
+                            {
+                                key: '0000001',
+                                title: '张三',
+                                _id: '0000001',
+                                name: '张三',
+                                type: 'member',
+                                disabled: true
+                            },
+                            {
+                                key: '0000003',
+                                title: '王五',
+                                _id: '0000001',
+                                name: '王五',
+                                type: 'member'
+                            }
+                        ]
                     },
                     {
                         key: '5d51537987367c0f92fe6fa8',